remove token file in backoffice
